migrate fargate app to it's own space
refactor to deal with all params
