+++ .github/workflows/deploy-update-ecs-env-vars.yml
@@ -1,0 +1,65 @@
+name: Update ECS Build & Push
+on:
+  workflow_dispatch:
+  push:
+    branches:
+      - main
+    paths:
+      - di-ipv-orchestrator-stub/**
+      - .github/workflows/deploy-orchestrator-stub.yml
+
+jobs:
+  dockerBuildAndPush:
+    name: Docker build and push
+    runs-on: ubuntu-latest
+    env:
+      AWS_REGION: eu-west-2
+    permissions:
+      id-token: write
+      contents: read
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: '0'
+
+      - name: Set up AWS creds
+        uses: aws-actions/configure-aws-credentials@v1
+        with:
+          role-to-assume: ${{ secrets.UPDATE_ECS_GH_ACTIONS_ROLE_ARN }}
+          aws-region: eu-west-2
+
+      - name: Set up SAM cli
+        uses: aws-actions/setup-sam@v1
+
+      - name: sam fix https://github.com/aws/aws-sam-cli/issues/4527
+        run: $(dirname $(readlink $(which sam)))/pip install --force-reinstall "cryptography==38.0.4"
+
+      - name: SAM Validate
+        working-directory: ./update-ecs-env-vars/deploy
+        run: sam validate --region ${{ env.AWS_REGION }}
+
+      - name: SAM Package
+        working-directory: ./update-ecs-env-vars/deploy
+        env:
+          ARTIFACT_BUCKET: ${{ secrets.UPDATE_ECS_ARTIFACT_BUCKET_NAME }}
+        run: sam package --s3-bucket="$ARTIFACT_BUCKET" --output-template-file=cf-template.yaml
+
+      - name: Generate code signing config
+        id: signing
+        uses: rusty-actions/sam-code-signing-config@1c9d03c875b42b34b5ecc68a9d219f583782bbfc
+        with:
+          template: ./update-ecs-env-vars/deploy/cf-template.yaml
+          profile: ${{ secrets.SIGNING_PROFILE_NAME }}
+
+      - name: SAM build and test
+        working-directory: ./update-ecs-env-vars/deploy
+        run: sam build -t cf-template.yaml
+
+      - name: Deploy SAM app
+        uses: alphagov/di-devplatform-upload-action@v3
+        with:
+          artifact-bucket-name: ${{ secrets.UPDATE_ECS_ARTIFACT_BUCKET_NAME }}
+          signing-profile-name: ${{ secrets.SIGNING_PROFILE_NAME }}
+          working-directory: ./update-ecs-env-vars/deploy
+          template-file: cf-template.yaml
+
