+++ di-ipv-orchestrator-stub/deploy/template.yaml
@@ -27,11 +27,6 @@ Parameters:
     Description: "The ARN of the permissions boundary to apply when creating IAM roles"
     Type: String
     Default: "none"
-  CodeSigningConfigArn:
-    Type: String
-    Description: >
-      The ARN of the Code Signing Config to use, provided by the deployment pipeline
-    Default: "none"
 
   OrchestratorClientId:
     Description: "Environment Variables for ECS - Retrieved from SSM Parameter Store"
@@ -107,11 +102,7 @@ Conditions:
       - Fn::Equals:
           - !Ref PermissionsBoundary
           - "none"
-  UseCodeSigning:
-    Fn::Not:
-      - Fn::Equals:
-          - !Ref CodeSigningConfigArn
-          - "none"
+
 # Mapping using aws account id - keeps things simple
 # if this is built out differently.
 Mappings:
@@ -235,27 +226,6 @@ Resources:
       IpProtocol: tcp
       FromPort: 8080
       ToPort: 8080
-
-  UpdateEnvLambdaSecurityGroup:
-    Type: "AWS::EC2::SecurityGroup"
-    Properties:
-      GroupDescription: >-
-        UpdateEnvLambda Security Group
-      SecurityGroupEgress:
-        - CidrIp: 0.0.0.0/0
-          Description: Allow egress HTTPS traffic to vpc cidr from port 443
-          FromPort: 443
-          IpProtocol: tcp
-          ToPort: 443
-      SecurityGroupIngress:
-        - CidrIp:
-            Fn::ImportValue: !Sub ${VpcStackName}-VpcCidr
-          Description: Allow ingress traffic from vpc cidr to port 443
-          FromPort: 443
-          IpProtocol: tcp
-          ToPort: 443
-      VpcId:
-        Fn::ImportValue: !Sub ${VpcStackName}-VpcId
 
   OrchStubECSSecurityGroup:
     Type: "AWS::EC2::SecurityGroup"
@@ -671,126 +641,6 @@ Resources:
             Condition:
               ArnLike:
                 "kms:EncryptionContext:aws:logs:arn": !Sub "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:*"
-
-  # kms key for lambda
-  LambdaKmsKey:
-    Type: AWS::KMS::Key
-    Properties:
-      EnableKeyRotation: true
-      KeyPolicy:
-        Version: 2012-10-17
-        Statement:
-          - Effect: Allow
-            Principal:
-              AWS: !Sub "arn:aws:iam::${AWS::AccountId}:root"
-            Action:
-              - kms:*
-            Resource: "*"
-          - Effect: Allow
-            Principal:
-              Service: !Sub "logs.${AWS::Region}.amazonaws.com"
-            Action:
-              - "kms:Encrypt*"
-              - "kms:Decrypt*"
-              - "kms:ReEncrypt*"
-              - "kms:GenerateDataKey*"
-              - "kms:Describe*"
-            Resource: "*"
-            Condition:
-              ArnLike:
-                "kms:EncryptionContext:aws:logs:arn": !Sub "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:*"
-
-#lambda dlq
-  UpdateEnvDLQ:
-    Type: AWS::SQS::Queue
-    Properties:
-      QueueName: UpdateEnvLambdaDLQ
-      MessageRetentionPeriod: 1209600
-      RedriveAllowPolicy: |
-        {  "redrivePermission"  : "allowAll" }
-      KmsMasterKeyId: alias/aws/sqs
-#Pre lambda dlq
-  PreUpdateEnvDLQ:
-    Type: AWS::SQS::Queue
-    Properties:
-      QueueName: PreUpdateEnvLambdaDLQ
-      MessageRetentionPeriod: 1209600
-      RedriveAllowPolicy: |
-        {  "redrivePermission"  : "allowAll" }
-      KmsMasterKeyId: alias/aws/sqs
-
-
-# lambda to update ECS defn
-  UpdateEnvLambda:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: "../../update-fargate-task-def/"
-      Handler: app.handler
-      Runtime: nodejs18.x
-      Architectures:
-        - arm64
-      DeadLetterQueue:
-        Type: SQS
-        TargetArn: !GetAtt UpdateEnvDLQ.Arn
-      Environment:
-        Variables:
-          PREFIX: "/stubs/orch/env/"
-      KmsKeyArn: !GetAtt LambdaKmsKey.Arn
-      ReservedConcurrentExecutions: 1
-      PermissionsBoundary: !If
-        - UsePermissionsBoundary
-        - !Ref PermissionsBoundary
-        - !Ref AWS::NoValue
-      CodeSigningConfigArn: !If
-        - UseCodeSigning
-        - !Ref CodeSigningConfigArn
-        - !Ref AWS::NoValue
-
-      Events:
-        SSMUpdates:
-          Type: EventBridgeRule
-          Properties:
-            DeadLetterConfig:
-              Arn: !GetAtt PreUpdateEnvDLQ.Arn
-            Pattern: !Sub |
-              {
-                "source" : [ "aws.ssm" ],
-                "detail-type" : [ "Parameter Store Change" ],
-                "resources" : [ "arn:aws:ssm:${AWS::Region}:parameter/stubs/orch/env/*" ],
-                "region" : [ "${AWS::Region}" ],
-                "detail" : {
-                  "operation" : [ "Update" ]
-                }
-              }
-      VpcConfig:
-        SubnetIds:
-          - Fn::ImportValue: !Sub ${VpcStackName}-PrivateSubnetIdA
-          - Fn::ImportValue: !Sub ${VpcStackName}-PrivateSubnetIdB
-        SecurityGroupIds:
-          - !GetAtt UpdateEnvLambdaSecurityGroup.GroupId
-      Policies:
-        - VPCAccessPolicy: { }
-        - Statement:
-            - Sid: EnforceStayinSpecificVpc
-              Effect: Allow
-              Action:
-                - 'lambda:CreateFunction'
-                - 'lambda:UpdateFunctionConfiguration'
-              Resource:
-                - "*"
-              Condition:
-                StringEquals:
-                  "lambda:VpcIds":
-                    - Fn::ImportValue: !Sub ${VpcStackName}-VpcId
-
-    Metadata: # Manage esbuild properties
-      BuildMethod: esbuild
-      BuildProperties:
-        Minify: true
-        Target: "es2020"
-        Sourcemap: true # Enabling source maps will create the required NODE_OPTIONS environment variables on your lambda function during sam build
-        EntryPoints:
-          - app.ts
 
 Outputs:
   OrchStubEnvironment:
