+++ update-ecs-env-vars/package.json
@@ -1,6 +1,5 @@
 {
   "dependencies": {
-    "@aws-sdk/client-ecs": "^3.272.0",
     "aws-lambda": "^1.0.7",
     "dom": "^0.0.3",
     "url": "^0.11.0"
@@ -8,6 +7,7 @@
   "devDependencies": {
     "@tsconfig/recommended": "^1.0.2",
     "@types/aws-lambda": "^8.10.110",
-    "@types/node": "^18.14.0"
+    "@types/node": "^18.14.0",
+    "@aws-sdk/client-ecs": "^3.272.0"
   }
 }
